chore: 폴더 구조화 및 스타일 변경
