+++ src/components/Header.tsx
@@ -11,6 +11,8 @@ export default function Header({ left, right }: { left: ReactNode; right?: React
 }
 
 const headerCss = css({
+  position: 'sticky',
+  top: 0,
   height: '64px',
   width: '430px',
   backgroundColor: 'black',
