+++ src/page/CartPage.tsx
@@ -1,7 +1,7 @@
 import { Link } from 'react-router';
 import Header from '../components/Header';
 import { css } from '@emotion/react';
-import CartItemList from '../components/CartItemList';
+import CartItemList from '../components/CartItemList/CartItemList';
 import { useApiContext } from '../contexts/ApiContext';
 import getCartItems from '../api/getCartItem';
 
@@ -19,7 +19,9 @@ function CartPage() {
       />
       <main css={layoutCss}>
         <h1 css={titleCss}>장바구니</h1>
-        {cartItems?.content.length !== 0 && <p>총 {cartItems?.content.length}개의 상품이 담겨 있습니다.</p>}
+        {cartItems?.content.length !== 0 && (
+          <p css={countCss}>총 {cartItems?.content.length}개의 상품이 담겨 있습니다.</p>
+        )}
         {cartItems?.content && <CartItemList cartItems={cartItems.content} />}
       </main>
     </>
@@ -44,4 +46,8 @@ const layoutCss = css({
 const titleCss = css({
   fontSize: '24px',
   fontWeight: '700'
+});
+
+const countCss = css({
+  marginBottom: '16px'
 });
